requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
alembic==1.9.3
    # via flask-migrate
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
cryptography==39.0.1
    # via pyopenssl
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-migrate
    #   flask-sqlalchemy
flask-login==0.6.2
    # via -r requirements.in
flask-migrate==4.0.4
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via
    #   -r requirements.in
    #   flask-migrate
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==6.0.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
mako==1.2.4
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
numpy==1.24.1
    # via pandas
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
oura==1.2.4
    # via -r requirements.in
pandas==1.5.3
    # via oura
psycopg2==2.9.5
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyopenssl==23.0.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.7.1
    # via pandas
requests==2.28.1
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via oura
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.3
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.4.0
    # via sqlalchemy
urllib3==1.26.13
    # via requests
werkzeug==2.2.2
    # via
    #   flask
    #   flask-login
zipp==3.11.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
